+++ taproot/index.js
@@ -1,3 +1,9 @@
 const attemptTaprootClaim = require('./attempt_taproot_claim');
+const taprootCoopTransaction = require('./taproot_coop_transaction');
+const taprootRefundTransaction = require('./taproot_refund_transaction');
 
-module.exports = {attemptTaprootClaim};
+module.exports = {
+  attemptTaprootClaim,
+  taprootCoopTransaction,
+  taprootRefundTransaction,
+};